+++ example_backend/examples/simple_button.rs
@@ -45,7 +45,11 @@ fn setup(mut commands: Commands, cli: Res<Cli>) -> Result<()> {
             let server = ExampleServer::new(port)?;
             commands.insert_resource(server);
 
-            commands.spawn((UiRoot, children![ToggleButton(false)]));
+            commands.spawn((
+                UiRoot,
+                Replicated,
+                children![(ToggleButton(false), Replicated)],
+            ));
             commands.spawn(Text::new("Server"));
         }
         Cli::Client { ip, port } => {
@@ -172,7 +176,6 @@ impl Default for Cli {
 /// this component is spawned) and the client (when the entity with this component is replicated).
 #[derive(Component, Serialize, Deserialize)]
 #[require(
-    Replicated,
     Node {
         width: Val::Percent(100.0),
         height: Val::Percent(100.0),
@@ -192,7 +195,6 @@ struct UiRoot;
 /// See <https://github.com/bevyengine/bevy/issues/19776> for more details.
 #[derive(Component, Serialize, Deserialize, Deref, DerefMut, Clone, Copy)]
 #[require(
-    Replicated,
     Button,
     Node {
         width: Val::Px(150.0),